code/weather_subscriber/requirements.txt
numpy==2.0.2
requests
python-dotenv
paho-mqtt==2.1.0


